refactor: InputStream 관리 개선 및 자동 자원 해제 적용
- try-with-resources 적용하여 InputStream 자동 close처리
- 예외 발생시 빈 리스트 반환하여 안정성 향상
